authorization-server/requirements.txt
anyio==3.5.0
APScheduler==3.6.3
asgiref==3.5.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
fastapi==0.75.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jwt==1.3.1
peewee==3.14.10
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
python-multipart==0.0.5
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tornado==6.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.17.6
